Revert "ZCS-8785:Fix for issue with data type range values for Mysql"
This reverts commit 11f1edf7981c266328c19d7310efd4e9ca03a094.

+++ store/src/java/com/zimbra/cs/imap/ImapSessionManager.java
@@ -772,7 +772,7 @@ final class ImapSessionManager {
     private String cacheKey(FolderStore folder, boolean active) {
         MailboxStore mbox = folder.getMailboxStore();
         int modseq = folder instanceof SearchFolderStore ? mbox.getLastChangeID() : folder.getImapMODSEQ();
-        long uvv = folder instanceof SearchFolderStore ? mbox.getLastChangeID() : ImapFolder.getUIDValidity(folder);
+        int uvv = folder instanceof SearchFolderStore ? mbox.getLastChangeID() : ImapFolder.getUIDValidity(folder);
         String acctId = null;
         try {
             acctId = mbox.getAccountId();
