+++ store/src/java/com/zimbra/cs/mailbox/MailServiceException.java
@@ -335,7 +335,7 @@ public class MailServiceException extends ServiceException {
         return new MailServiceException("WaitSet not found: " + id, NO_SUCH_WAITSET, SENDERS_FAULT, new Argument(ID, id, Argument.Type.STR));
     }
 
-    public static MailServiceException NO_SUCH_BLOB(int mboxId, int itemId, long revision) {
+    public static MailServiceException NO_SUCH_BLOB(int mboxId, int itemId, int revision) {
         return new MailServiceException("No such blob: mailbox=" + mboxId + ", item=" + itemId + ", change=" + revision, NO_SUCH_BLOB, SENDERS_FAULT,
             new Argument(ITEM_ID, itemId, Argument.Type.IID), new Argument(REVISION, revision, Argument.Type.NUM));
     }
